backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
dnspython==2.6.1
pydantic==2.7.1
